Added new type for search options to faciliate ::S commands. Fixed lint warning.

+++ src/file.c
@@ -525,5 +525,5 @@ void set_last(const char *name)
         throw(E_MAX);                   // Internal program limit reached
     }
 
-    last_len = (uint_t)nbytes;
+    last_len = (uint_t)(uint)nbytes;
 }
